+++ src/components/componentMaintenanceCalories/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-maintenance-calories
 
+## 0.0.2
+
+### Patch Changes
+
+- adding age in to users personal data to calculate maintenance calores
+
 ## 0.0.1
 
 ### Patch Changes